move table in domain overview
