+++ src/domains/overview/tableModel/presentation/TableModelController.tsx
@@ -1,23 +1,23 @@
 /* eslint-disable react/display-name */
 import React, { FunctionComponent, useEffect, useRef, useState } from "react";
 import { useSelector } from "react-redux";
-import { ChainlinkCTO } from "../../../dataAccess/access/cto/ChainlinkCTO";
-import { SequenceCTO } from "../../../dataAccess/access/cto/SequenceCTO";
-import { SequenceStepCTO } from "../../../dataAccess/access/cto/SequenceStepCTO";
-import { ActionTO } from "../../../dataAccess/access/to/ActionTO";
-import { ChainDecisionTO } from "../../../dataAccess/access/to/ChainDecisionTO";
-import { ChainTO } from "../../../dataAccess/access/to/ChainTO";
-import { ConditionTO } from "../../../dataAccess/access/to/ConditionTO";
-import { DataSetupTO } from "../../../dataAccess/access/to/DataSetupTO";
-import { DecisionTO } from "../../../dataAccess/access/to/DecisionTO";
-import { SequenceTO } from "../../../dataAccess/access/to/SequenceTO";
-import { CalcChain } from "../../../services/SequenceChainService";
-import { CalculatedStep } from "../../../services/SequenceService";
-import { editSelectors, Mode } from "../../../slices/EditSlice";
-import { masterDataSelectors } from "../../../slices/MasterDataSlice";
-import { sequenceModelSelectors } from "../../../slices/SequenceModelSlice";
-import { DavitUtil } from "../../../utils/DavitUtil";
-import { DavitTable } from "../../common/fragments/DavitTable";
+import { DavitTable } from "../../../../components/common/fragments/DavitTable";
+import { ChainlinkCTO } from "../../../../dataAccess/access/cto/ChainlinkCTO";
+import { SequenceCTO } from "../../../../dataAccess/access/cto/SequenceCTO";
+import { SequenceStepCTO } from "../../../../dataAccess/access/cto/SequenceStepCTO";
+import { ActionTO } from "../../../../dataAccess/access/to/ActionTO";
+import { ChainDecisionTO } from "../../../../dataAccess/access/to/ChainDecisionTO";
+import { ChainTO } from "../../../../dataAccess/access/to/ChainTO";
+import { ConditionTO } from "../../../../dataAccess/access/to/ConditionTO";
+import { DataSetupTO } from "../../../../dataAccess/access/to/DataSetupTO";
+import { DecisionTO } from "../../../../dataAccess/access/to/DecisionTO";
+import { SequenceTO } from "../../../../dataAccess/access/to/SequenceTO";
+import { CalcChain } from "../../../../services/SequenceChainService";
+import { CalculatedStep } from "../../../../services/SequenceService";
+import { editSelectors, Mode } from "../../../../slices/EditSlice";
+import { masterDataSelectors } from "../../../../slices/MasterDataSlice";
+import { sequenceModelSelectors } from "../../../../slices/SequenceModelSlice";
+import { DavitUtil } from "../../../../utils/DavitUtil";
 import { TabPanel } from "../fragments/TabPanel";
 import { useGetCalcErrorActionsTableData } from "../tables/calculated/CalcErrorActions";
 import { useGetCalcLinkTableData } from "../tables/calculated/CalcLink";
