+++ src/domains/overview/tableModel/tables/model/ModelDataSetup.ts
@@ -1,9 +1,9 @@
 import { useDispatch } from "react-redux";
-import { DataSetupTO } from "../../../../dataAccess/access/to/DataSetupTO";
-import { EditActions } from "../../../../slices/EditSlice";
-import { SequenceModelActions } from "../../../../slices/SequenceModelSlice";
-import { DavitTableRowData } from "../../../common/fragments/DavitTable";
-import { DavitIcons } from "../../../common/IconSet";
+import { DavitTableRowData } from "../../../../../components/common/fragments/DavitTable";
+import { DavitIcons } from "../../../../../components/common/IconSet";
+import { DataSetupTO } from "../../../../../dataAccess/access/to/DataSetupTO";
+import { EditActions } from "../../../../../slices/EditSlice";
+import { SequenceModelActions } from "../../../../../slices/SequenceModelSlice";
 
 export const useGetDataSetupTableData = (dataSetups: DataSetupTO[]) => {
     const dispatch = useDispatch();
