+++ src/domains/overview/tableModel/tables/model/ModelSequence.ts
@@ -1,9 +1,9 @@
 import { useDispatch } from "react-redux";
-import { SequenceTO } from "../../../../dataAccess/access/to/SequenceTO";
-import { EditActions } from "../../../../slices/EditSlice";
-import { SequenceModelActions } from "../../../../slices/SequenceModelSlice";
-import { DavitTableAction, DavitTableRowData } from "../../../common/fragments/DavitTable";
-import { DavitIcons } from "../../../common/IconSet";
+import { DavitTableAction, DavitTableRowData } from "../../../../../components/common/fragments/DavitTable";
+import { DavitIcons } from "../../../../../components/common/IconSet";
+import { SequenceTO } from "../../../../../dataAccess/access/to/SequenceTO";
+import { EditActions } from "../../../../../slices/EditSlice";
+import { SequenceModelActions } from "../../../../../slices/SequenceModelSlice";
 
 export const useGetSequenceModelsTableBody = (sequences: SequenceTO[]) => {
     const dispatch = useDispatch();
