+++ src/domains/overview/tableModel/util/TableUtil.ts
@@ -1,6 +1,6 @@
-import { ChainlinkCTO } from "../../../dataAccess/access/cto/ChainlinkCTO";
-import { ChainDecisionTO } from "../../../dataAccess/access/to/ChainDecisionTO";
-import { GoToChain, GoToTypesChain, IntermediateChain } from "../../../dataAccess/access/types/GoToTypeChain";
+import { ChainlinkCTO } from "../../../../dataAccess/access/cto/ChainlinkCTO";
+import { ChainDecisionTO } from "../../../../dataAccess/access/to/ChainDecisionTO";
+import { GoToChain, GoToTypesChain, IntermediateChain } from "../../../../dataAccess/access/types/GoToTypeChain";
 
 export function getChainGotoName(
     goto: GoToChain,
