+++ src/pages/Davit.tsx
@@ -8,10 +8,10 @@ import "../app/css/Davit.css";
 import { ErrorNotification } from "../components/common/fragments/ErrorNotification";
 import { ControlPanelController } from "../components/controllPanel/presentation/ControlPanelController";
 import { SidePanelController } from "../components/sidePanel/SidePanelController";
-import { TableModelController } from "../components/tableModel/presentation/TableModelController";
 import { ActorModelController } from "../domains/actor/ActorModelController";
 import { DataModelController } from "../domains/datamodel/DataModelController";
 import { FlowChartController } from "../domains/overview/flowChartModel/FlowChartController";
+import { TableModelController } from "../domains/overview/tableModel/presentation/TableModelController";
 import { GlobalActions } from "../slices/GlobalSlice";
 import { MasterDataActions } from "../slices/MasterDataSlice";
